requirements.txt
pandas
scikit-learn
joblib
PyMuPDF
lightgbm
google-generativeai
chromadb
flask
